refactor(group-quiz): extract invite-share into useGroupQuizShare
Move canNativeShare detection + handleCopy/handleNativeShare verbatim
into _hooks/use-group-quiz-share.ts (owns its own GroupQuiz.session
translations, takes the session). Replaced the now-unused
nativeShareOrCopy import with the hook. Behavior-identical;
page.tsx 1314->1288 LOC.
